Update features
I've done a lot of stuff, so I checked off a lot of boxes. I still need to add withdraw stuff to this list but the current state of matching right now is fixing bugs.
Matching only happens when an order comes in, but I'm assuming lots of orders will come in and theoretically every order gets matched as fast as possible.