got db to compile, removed stuff
Next step is to refactor so we're using our new CoinList. I really don't want to have to deal with coin-specific stuff and this should be compatible with as many coins as supported on lit. Once that's done, the code will be very easy to change into whatever I need it to do for the channel pushing stuff. I already confirmed that it's able to get events.