Deleted dextechsurvey.md
Now that this repo has a bunch of working code, it might be a good idea to
create some other document to outline and analyze all of this, such as a
wiki. It's also important to stay objective, and dextechsurvey.md
wasn't super objective, even if there was some good analysis.